create types file
